requirements.txt
Flask
Flask-Login
Flask-Session
Flask_SocketIO
itsdangerous
Jinja2
MarkupSafe
python-engineio
python-socketio
six
Werkzeug
eventlet
